Added TIMERFMS and reworked TIMERF ticking system.
- Added: TIMERFMS, which has the same syntax and usage as TIMERF but accepts a timeout in milliseconds.
- Changed: TIMERF(MS) now uses the same (faster) ticking system that every other "Timed Object" uses. This also removes the real-world ticking time difference between an item's timer and a timerf.
    Now an item with TIMER=10 and TIMERF with delay 10 will immediately be executed one after another, without delays caused by the different ticking systems used before.
- Fixed: Server crash when a TIMERF executed TIMERF STOP/CLEAR, altering the timed functions list while one of them was running (Issue #516).
- Removed: DEBUGF_SCRIPTS log messages for the following instructions: ADDCLILOC, REPLACECLILOC, REMOVECLILOC, STRSUB, MULDIV.

+++ src/game/items/CItem.cpp
@@ -4155,7 +4155,7 @@ void CItem::ConvertBolttoCloth()
 		if ( pBaseDef == nullptr )
 			continue;
 
-        uint iTotalAmount = iOutAmount * pDefCloth->m_BaseResources[i].GetResQty();
+        int64 iTotalAmount = iOutAmount * pDefCloth->m_BaseResources[i].GetResQty();
         
         CItem* pItemNew = nullptr;
         while (iTotalAmount > 0)
